requirements.txt
pandas
httpx
langchain
crewai
langchain
langchain-core
langchain-community
langchain-openai
langchain-experimental
litellm
fastapi
uvicorn
fastapi
uvicorn
sqlite-utils
mysql-connector-python
python-dotenv    
pymysql
sqlalchemy
crewai-tools
fastapi
uvicorn
streamlit
pydantic
requests

